Started block user feature. Improvements in admin interface.

+++ app/Http/Controllers/UserController.php
@@ -69,10 +69,13 @@ class UserController extends Controller
             $communities = Community::where('id_owner', '=', $user_id)->orderBy('name', 'asc')->get();
 
             $request_status = null;
+            $follows = null;
+            $blocking= null;
+            $blocked= null;
 
             if ($this_user !== null) {
-                $condition = ['id_receiver' => $user_id, "id_sender" => $this_user->id];
-                $request = Request::where($condition)->get();
+                $requestCondition = ['id_receiver' => $user_id, "id_sender" => $this_user->id];
+                $request = Request::where($requestCondition)->get();
                 if (sizeof($request) > 0) {
                     //dd($request);
                     for ($i = 0; $i < sizeof($request); $i++) {
@@ -85,11 +88,37 @@ class UserController extends Controller
                         }
                     }
                 }
+
+                $followCondition = ['id_follower' => $this_user->id, 'id_followed' => $user_id];
+                $follow_row = DB::table('follow_user')->where($followCondition)->get(); 
+                if(sizeof($follow_row) > 0){
+                    $follows = true;
+                } else {
+                    $follows = false;
+                }
+
+                $blockingCondition = ['blocked_user' => $user_id, 'blocker_user' => $this_user->id];
+                $block_row = DB::table('block_user')->where($blockingCondition)->first(); 
+                if($block_row !== null){
+                    $blocking = true;
+                } else {
+                    $blocking = false;
+                }
+
+                $blockedCondition = ['blocked_user' => $this_user->id, 'blocker_user' => $user_id];
+                $block_row = DB::table('block_user')->where($blockedCondition)->first(); 
+                if($block_row !== null){
+                    $blocked = true;
+                } else {
+                    $blocked = false;
+                }
+
             }
+            //dd($blocking, $blocked);
             // $follow_request = DB::table('follow_user')->where('id_followed', '=', $user_id)->get() !== null;
 
             //$comments = DB::table('comment')->where('id_post', '=', $id)->orderBy('time_stamp', 'desc')->get();
-            return view('pages.myProfile', ['other_user' => $member_user, 'age' => $age, 'nPosts' => $postN, 'posts' => $posts, 'communities' => $communities, 'user' => $this_user, 'follow_status' => $request_status]);
+            return view('pages.myProfile', ['other_user' => $member_user, 'age' => $age, 'nPosts' => $postN, 'posts' => $posts, 'communities' => $communities, 'user' => $this_user, 'follow_status' => $request_status, 'follows' => $follows, 'isBlocked' => $blocked, 'isBlocking'=> $blocking]);
         }
         abort(404);
     }
@@ -302,10 +331,30 @@ class UserController extends Controller
                 ->join('follow_request', 'follow_request.id', '=', 'request.id')
                 ->where($condition)
                 ->first();
+           if($request != null){
             DB::delete('delete from notification where id_request = ?', [$request->id]);
             DB::delete('delete from follow_request where id = ?', [$request->id]);
             $request = Request::find($request->id);
             $request->delete();
+           }
+           DB::delete('delete from follow_user where id_followed = ? and id_follower = ?', [$user_id, $user->id]);
         });
+    }
+
+    public function block($user_id)
+    {
+        $user = Auth::user();
+        DB::transaction(function () use ($user_id, $user) {
+            DB::insert('insert into block_user (blocked_user, blocker_user) values (?, ?)', [$user_id, $user->id]);
+            DB::delete('delete from request where id_receiver = ? and id_sender = ?', [$user->id, $user_id]);
+            DB::delete('delete from request where id_receiver = ? and id_sender = ?', [$user_id, $user->id]);
+        });
+    }
+
+    public function unblock($user_id)
+    {
+        $user = Auth::user();
+        DB::delete('delete from block_user where blocked_user = ? and blocker_user = ?', [$user_id, $user->id]);
+        
     }
 }
